Moved the appliance name and type from DeviceDetail fragment to ActionBar.
