+++ Nuton/src/main/res/layout/actionbar_device_detail.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/action_bar_device_detail"
+    android:orientation="horizontal"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <Spinner
+        android:id="@+id/action_bar_device_type"
+        android:drawSelectorOnTop="true"
+        android:layout_height="match_parent"
+        android:background="@color/background"
+        android:layout_width="wrap_content"/>
+
+    <EditText
+        android:id="@+id/action_bar_device_name"
+        android:inputType="textCapWords"
+        android:layout_height="match_parent"
+        android:layout_width="wrap_content"
+        android:hint="@string/actionbar_device_name_hint"
+        android:text=""
+        android:textColor="@android:color/white"
+        android:background="@android:color/transparent"/>
+
+</LinearLayout>
