+++ Nuton/src/main/res/layout/fragment_device_detail.xml
@@ -9,23 +9,13 @@
         android:layout_height="fill_parent">
 
         <GridLayout
-            android:rowCount="6"
+            android:rowCount="7"
             android:columnCount="2"
             android:layout_margin="15dp"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:descendantFocusability="beforeDescendants"
             android:focusableInTouchMode="true">
-
-            <EditText
-                android:id="@+id/dd_name"
-                android:inputType="textCapWords"
-                android:layout_row="0"
-                android:layout_column="0"
-                android:textSize="26sp"
-                android:text=""
-                android:textColor="@android:color/white"
-                android:background="@android:color/transparent"/>
 
             <at.markushi.ui.CircleButton
                 android:id="@+id/dd_btn_update_firmware"
@@ -48,17 +38,26 @@
                 android:layout_height="wrap_content"
                 android:layout_gravity="end"/>
 
-            <Spinner
-                android:id="@+id/dd_lst_connected_device"
-                android:spinnerMode="dialog"
-                android:drawSelectorOnTop="true"
+            <View
                 android:layout_row="2"
                 android:layout_column="0"
-                android:layout_columnSpan="2"/>
+                android:layout_columnSpan="2"
+                android:layout_width="fill_parent"
+                android:layout_height="1dp"
+                android:background="@android:color/darker_gray"/>
+
+            <CheckBox
+                android:layout_row="3"
+                android:layout_column="0"
+                android:text="Turn on/off automatically"/>
+            <CheckBox
+                android:layout_row="4"
+                android:layout_column="0"
+                android:text="Enable remote control"/>
 
             <com.github.mikephil.charting.charts.BarChart
                 android:id="@+id/ddl_chart"
-                android:layout_row="3"
+                android:layout_row="5"
                 android:layout_column="0"
                 android:layout_columnSpan="2"
                 android:minHeight="400dp"
@@ -67,7 +66,7 @@
 
             <Button
                 android:id="@+id/dd_btn_test_data"
-                android:layout_row="4"
+                android:layout_row="6"
                 android:layout_column="0"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
