+++ Nuton/src/main/res/layout/spinner_appliance_type.xml
@@ -6,19 +6,6 @@
     <ImageView
         android:id="@+id/spinner_appliance_image"
         android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginEnd="10dp" />
+        android:layout_height="40dp"/>
 
-    <TextView
-        android:id="@+id/spinner_appliance_type_name"
-        android:textSize="16sp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_toEndOf="@+id/spinner_appliance_image"
-        android:ellipsize="marquee"
-        android:gravity="fill_vertical"
-        android:singleLine="true"
-        android:layout_marginEnd="10dp"/>
 </RelativeLayout>
