+++ Nuton/src/main/res/values/strings.xml
@@ -11,6 +11,8 @@
 
     <string name="main_action_bar_settings">Settings</string>
 
+    <string name="actionbar_device_name_hint">Appliance Name</string>
+
     <string name="ble_not_supported">BLE is not supported by this Android version.</string>
     <string name="error_bluetooth_not_supported">This phone does not have support for BLE hardware.</string>
     <string name="dl_btn_connect">Connect</string>
